+++ src/plugins/calc.cpp
