+++ src/plugins/cg.cpp
